requirements.txt
anyio==4.0.0
certifi==2023.7.22
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.1
Pillow==10.1.0
ply==3.11
pygame==2.5.0
PyQt5-Qt5==5.15.15
PyQt5-sip==12.13.0
python-telegram-bot==20.5
sip==6.7.9
sniffio==1.3.0
tk==0.1.0